MERGE: -c 61864 /branches/Platform-6.1 -> /trunk/Platform
Fix build break.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@61927 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ samba-interop/wbclient/lsa_wbc_sid.c
@@ -534,7 +534,7 @@ wbcErr wbcLookupSids(
     PLSA_SID_INFO pNameList = NULL;
     int index = 0;
     int domainIndex = 0;
-    int numDomains = 0;
+    size_t numDomains = 0;
     struct wbcTranslatedName *pNames = NULL;
     struct wbcDomainInfo *pDomains = NULL;
